config with embed_dim changes and ToEndMemberConv - Modify unmixing collection dataset config to enable caching - Adjust training parameters and loss configurations for both models - Enhance visualization functions in unmixing trainer - Support EMA evaluation and improved endmember visualization - Refactor main training script to handle multiple model types dynamically
feat(config): add new tokenizer GAN configuration files

- Introduce uni_cosmos_f8c16p1_nested tokenizer config
- Add unicosmos_tokenizer_psd_f8c16p1.yaml for PSD quantizer training
- Configure REPA and VF losses with DINO feature settings
- Set up optimizers and schedulers for tokenizer and discriminator

feat(submodule): add SphereAR submodule

- Register SphereAR repository as git submodule under src/stage1/VAEs
- Point to guolinke/SphereAR.git repository

fix(trainer): correct validation dataloader reloading logic

- Fix bug where val_dataloader was not properly reloaded during validation
- Ensure generator-type dataloaders are correctly handled in validation loop

refactor(unmixing): improve type hints and model initialization

- Add proper type annotations to setup_unmixing_model method
- Explicitly cast model to DownstreamModelTokenizerWrapper
- Improve assertion messages and code clarity in trainer methods

docs(trainer): enhance docstrings for visualization functions

- Add comprehensive docstrings to reconstruction visualization methods
- Clarify parameters and functionality of unmixing result visualizations

chore(config): adjust unmixing model hyperparameters

- Reduce batch size for Vitamin unmixing model
- Modify learning rates and optimizer settings
- Tune loss weights and scheduling parameters
- Update validation steps and duration settingsfeat(unmixing): add ResNet-based unmixing model and configs

- Add new ResNet unmixing model configuration with tokenizer wrapper
- Include full YAML config for ResNet-based unmixing pipeline
- Update Vitamin model

+++ src/stage2/unmixing/metrics/basic.py
@@ -314,8 +314,8 @@ class UnmixingMetrics(torch.nn.Module):
             axes = axes.reshape(2, 1)
 
         # Convert to numpy for plotting
-        endmembers_gt_np = endmembers_gt.detach().cpu().numpy()
-        ordered_endmembers_np = ordered_endmembers.detach().cpu().numpy()
+        endmembers_gt_np = endmembers_gt.float().detach().cpu().numpy()
+        ordered_endmembers_np = ordered_endmembers.float().detach().cpu().numpy()
 
         # Plot endmembers
         for i in range(num_endmembers):
